pyproject.toml
[tool.poetry]
name = "ideal-pancake"
version = "0.1.0"
description = ""
authors = ["s0lvang <august.s.solvang@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.19.2"
pandas = "^1.1.2"
matplotlib = "^3.3.2"
pyEDFlib = "^0.1.18"
sklearn = "^0.0"
statsmodels = "^0.12.0"
arch = "^4.15"
xlrd = "^1.2.0"
scikit-learn = "^0.23.2"
scikit-plot = "*"
tensorflow = "2.2.0"
google-api-python-client = "*"
six = "^1.13.0"
pandas-gbq="0.14.0"
google-cloud-storage = "^1.32.0"
tsfresh = "^0.17.0"
numba = "^0.51.2"
scipy = {version = "1.4.1", allow-prereleases = true}
google-cloud-core = "^1.4.3"
cloudml-hypertune = "^0.1.0-alpha.6"
Keras = "^2.4.3"
opencv-python = "^4.4.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
black = {version="*", allow-prereleases=true}

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[flake8]
ignore = "E111"
